signed long -> signed int
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@3650 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ Binaries/dcmtk-source/dcmdata/dcddirif.cc
@@ -714,7 +714,7 @@ static void addConceptModContentItems(DcmDirectoryRecord *record,
     if ((record != NULL) && (dataset != NULL))
     {
         OFString tmpString;
-        signed long i = 0;
+        signed int i = 0;
         DcmItem *ditem = NULL;
         /* create new ContentSequence */
         DcmSequenceOfItems *newSeq = new DcmSequenceOfItems(DCM_ContentSequence);
